+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/action/swf3/ActionGoToLabel.java
@@ -98,7 +98,7 @@ public class ActionGoToLabel extends Action {
      */
     @Override
     protected int getContentBytesLength() {
-        return Utf8Helper.getBytesLength(label) + 1;
+        return Utf8Helper.getBytesLength(label, charset) + 1;
     }
 
     /**